Redisign CASADemo to use Command pattern. New C# test for RSProxy. Doxygen documention corrections
work-items: #17456, #17470, #19695
git-tfs-id: [https://sede-pt-ce2.visualstudio.com/DefaultCollection]$/Basin Modeling/IBS/Trunk;C46189

+++ development/libraries/casaAPI/src/RSProxyImpl.C
@@ -129,7 +129,7 @@ ErrorHandler::ReturnCode RSProxyImpl::calculateRSProxy( const std::vector<const 
    {
       m_collection->calculate( targetCollection
                              , validCase2Obs 
-                             , m_rsOrder
+                             , static_cast<unsigned int>( m_rsOrder )
                              , m_autosearch
                              , m_targedR2
                              , m_confLevel
